refactor(roles): split assignment modal, standardize typography, and move types to central file
