+++ src/app/(app-shell)/(management)/settings/roles/page.tsx
@@ -7,7 +7,6 @@ import { db } from '@/db';
 import { departments, users } from '@/db/schema';
 import type { UserRole } from '@/types/auth';
 
-import { RolesAddUserButton } from '../../../../../components/features/roles/roles-add-user-button';
 import { RolesManagementTable } from '../../../../../components/features/roles/roles-management-table';
 
 const ROLE_CONFIG: Array<{
@@ -167,18 +166,13 @@ export default async function RolesPage({ searchParams }: RolesPageProps) {
           <p className={`max-w-175 text-muted-foreground ${textSmRegularClass}`}>
             {selectedRoleInfo.description}
           </p>
-
-          <RolesAddUserButton
-            selectedRole={selectedRole}
-            mappedUsers={usersInRole}
-            currentUserId={currentUser.id}
-          />
         </div>
 
         <RolesManagementTable
           users={usersInRole}
           roleLabel={selectedRoleInfo.name}
           currentUserId={currentUser.id}
+          selectedRole={selectedRole}
         />
       </section>
     </div>
