--- src/components/features/roles/user-role-assignment-modal.tsx
@@ -1,526 +1,0 @@
-"use client"
-
-import { useEffect, useMemo, useState } from "react"
-import { CirclePlus, Info, Search, Trash2, X } from "lucide-react"
-
-import { assignUserRole, assignUsersRoleBulk, searchUsers } from "@/actions/roles"
-import { Avatar, AvatarFallback } from "@/components/ui/avatar"
-import { Button } from "@/components/ui/button"
-import { Checkbox } from "@/components/ui/checkbox"
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import { Input } from "@/components/ui/input"
-import type { UserRole } from "@/types/auth"
-
-export type RoleUser = {
-  id: string
-  name: string
-  email: string
-  department: string
-  role: UserRole
-}
-
-interface UserRoleAssignmentModalProps {
-  isOpen: boolean
-  onOpenChange: (open: boolean) => void
-  user: RoleUser | null
-  mode?: "edit" | "add"
-  defaultRole?: UserRole
-  mappedUsers?: RoleUser[]
-  onUpdated?: () => void
-  currentUserId: string
-}
-
-const ROLE_OPTIONS: Array<{ value: UserRole; label: string }> = [
-  { value: "GlobalAdmin", label: "Global Admin" },
-  { value: "ITOperator", label: "IT Operator" },
-  { value: "FinanceAuditor", label: "Finance Auditor" },
-  { value: "Employee", label: "Employee" },
-]
-
-const ROLE_ASSIGNMENT_LABELS: Record<UserRole, string> = {
-  GlobalAdmin: "Global Admin",
-  ITOperator: "IT Operations",
-  FinanceAuditor: "Auditor",
-  Employee: "Employee",
-}
-
-function getInitials(name: string) {
-  return name
-    .split(" ")
-    .map((part) => part[0] ?? "")
-    .join("")
-    .slice(0, 2)
-    .toUpperCase()
-}
-
-export function UserRoleAssignmentModal({
-  isOpen,
-  onOpenChange,
-  user,
-  mode = "edit",
-  defaultRole = "Employee",
-  mappedUsers = [],
-  onUpdated,
-  currentUserId,
-}: UserRoleAssignmentModalProps) {
-  const [selectedRole, setSelectedRole] = useState<UserRole>("Employee")
-  const [searchQuery, setSearchQuery] = useState("")
-  const [searchResults, setSearchResults] = useState<RoleUser[]>([])
-  const [isSearching, setIsSearching] = useState(false)
-  const [searchError, setSearchError] = useState<string | null>(null)
-  const [hideUsersAlreadyInRole, setHideUsersAlreadyInRole] = useState(false)
-  const [mappedSelection, setMappedSelection] = useState<RoleUser[]>([])
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-
-  const roleLabelForAddMode = ROLE_ASSIGNMENT_LABELS[defaultRole]
-
-  const normalizedQuery = searchQuery.trim()
-
-  const mappedIdSet = useMemo(
-    () => new Set(mappedSelection.map((roleUser) => roleUser.id)),
-    [mappedSelection]
-  )
-
-  const alreadyAssignedIdSet = useMemo(
-    () => new Set(mappedUsers.map((u) => u.id)),
-    [mappedUsers]
-  )
-
-  // Directory results exclude already-staged users and optionally hide already-mapped users.
-  const directoryResults = useMemo(() => {
-    if (mode !== "add" || !normalizedQuery) {
-      return []
-    }
-
-    return searchResults
-      .filter((directoryUser) => {
-        // Hide the current active user so they can't lock themselves out
-        if (directoryUser.id === currentUserId) {
-          return false
-        }
-
-        // Hide users that are already selected in the staging area below
-        if (mappedIdSet.has(directoryUser.id)) {
-          return false
-        }
-
-        // Accurately hide users already assigned to the target role
-        if (hideUsersAlreadyInRole) {
-          if (directoryUser.role === defaultRole || alreadyAssignedIdSet.has(directoryUser.id)) {
-            return false
-          }
-        }
-
-        return true
-      })
-      .slice(0, 10)
-  }, [currentUserId, hideUsersAlreadyInRole, mappedIdSet, alreadyAssignedIdSet, mode, normalizedQuery, searchResults, defaultRole])
-
-  const activeUser = mode === "edit" ? user : null
-
-  const addUserToSelection = (directoryUser: RoleUser) => {
-    setMappedSelection((currentSelection) => {
-      if (currentSelection.some((selection) => selection.id === directoryUser.id)) {
-        return currentSelection
-      }
-
-      return [...currentSelection, directoryUser]
-    })
-  }
-
-  const removeUserFromSelection = (roleUserId: string) => {
-    setMappedSelection((currentSelection) =>
-      currentSelection.filter((selection) => selection.id !== roleUserId)
-    )
-  }
-
-  const [prevIsOpen, setPrevIsOpen] = useState(isOpen)
-  const [prevMode, setPrevMode] = useState(mode)
-  const [prevUser, setPrevUser] = useState(user)
-
-  if (isOpen && !prevIsOpen) {
-    setPrevIsOpen(true)
-    setError(null)
-    setIsSubmitting(false)
-    if (mode === "add") {
-      setSelectedRole(defaultRole)
-      setSearchQuery("")
-      setSearchResults([])
-      setIsSearching(false)
-      setSearchError(null)
-      setHideUsersAlreadyInRole(false)
-      setMappedSelection([])
-    } else if (user) {
-      setSelectedRole(user.role)
-    }
-  } else if (!isOpen && prevIsOpen) {
-    setPrevIsOpen(false)
-    setIsSubmitting(false)
-    setError(null)
-    setSelectedRole(mode === "add" ? defaultRole : (user?.role ?? "Employee"))
-    if (mode === "add") {
-      setSearchQuery("")
-      setSearchResults([])
-      setIsSearching(false)
-      setSearchError(null)
-      setHideUsersAlreadyInRole(false)
-      setMappedSelection([])
-    }
-  }
-
-  if (mode !== prevMode || user !== prevUser) {
-    setPrevMode(mode)
-    setPrevUser(user)
-    if (mode === "edit" && user) {
-      setSelectedRole(user.role)
-    } else if (mode === "add") {
-      setSelectedRole(defaultRole)
-    }
-  }
-
-  useEffect(() => {
-    if (!isOpen || mode !== "add") {
-      return
-    }
-
-    if (!normalizedQuery) {
-      return
-    }
-
-    let isCancelled = false
-    const searchDebounce = setTimeout(async () => {
-      setIsSearching(true)
-      setSearchError(null)
-
-      try {
-        const results = await searchUsers(normalizedQuery)
-        if (isCancelled) {
-          return
-        }
-
-        setSearchResults(results)
-      } catch (caughtError) {
-        if (isCancelled) {
-          return
-        }
-
-        setSearchResults([])
-        setSearchError(
-          caughtError instanceof Error
-            ? caughtError.message
-            : "Failed to search users."
-        )
-      } finally {
-        if (!isCancelled) {
-          setIsSearching(false)
-        }
-      }
-    }, 250)
-
-    return () => {
-      isCancelled = true
-      clearTimeout(searchDebounce)
-    }
-  }, [isOpen, mode, normalizedQuery])
-
-  // Handles two submit flows: bulk mapping in add mode and single-user update in edit mode.
-  const handleSubmit = async () => {
-    setIsSubmitting(true)
-    setError(null)
-
-    try {
-      if (mode === "add") {
-        if (mappedSelection.length === 0) {
-          setError("Select at least one user to map.")
-          return
-        }
-
-        const result = await assignUsersRoleBulk(
-          mappedSelection.map((selection) => selection.id),
-          defaultRole
-        )
-
-        if (!result.success) {
-          setError(result.error ?? "Failed to assign selected users.")
-          return
-        }
-
-        onOpenChange(false)
-        onUpdated?.()
-        return
-      }
-
-      if (!activeUser) {
-        setError("Select a user to update.")
-        return
-      }
-
-      const result = await assignUserRole(activeUser.id, selectedRole)
-
-      if (!result.success) {
-        setError(result.error ?? "Failed to update role.")
-        return
-      }
-
-      onOpenChange(false)
-      onUpdated?.()
-    } catch (caughtError) {
-      setError(
-        caughtError instanceof Error
-          ? caughtError.message
-          : mode === "add"
-            ? "Failed to assign selected users."
-            : "Failed to update role. Please try again."
-      )
-    } finally {
-      setIsSubmitting(false)
-    }
-  }
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-190 p-0 overflow-hidden border-none shadow-2xl [&>button]:hidden">
-        <DialogTitle className="sr-only">
-          {mode === "add" ? `Assign Users to ${roleLabelForAddMode}` : "Change User Role"}
-        </DialogTitle>
-        <DialogDescription className="sr-only">
-          {mode === "add"
-            ? `Search and map users to ${roleLabelForAddMode}.`
-            : user
-              ? `Update permissions for ${user.name} (${user.email}).`
-              : "Select a user to update their role."}
-        </DialogDescription>
-
-        {mode === "add" ? (
-          <div className="p-4 sm:p-5">
-            <div className="flex items-start justify-between gap-3">
-              <div className="flex items-center gap-2">
-                <Info className="h-5 w-5 text-muted-foreground" />
-                <h2 className="text-lg font-bold text-foreground">
-                  Assign Users to {roleLabelForAddMode}
-                </h2>
-              </div>
-
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-xs"
-                className="-mr-1 -mt-1 text-muted-foreground hover:bg-muted"
-                onClick={() => onOpenChange(false)}
-              >
-                <X className="h-3.5 w-3.5" />
-              </Button>
-            </div>
-
-            <div className="mt-3 flex items-center gap-2">
-              <Checkbox
-                id="hide-already-mapped"
-                checked={hideUsersAlreadyInRole}
-                onCheckedChange={(checked) => setHideUsersAlreadyInRole(checked === true)}
-                disabled={isSubmitting}
-                className="border-border"
-              />
-              <label htmlFor="hide-already-mapped" className="text-sm font-medium text-foreground">
-                Hide users already in this role
-              </label>
-            </div>
-
-            <div className="mt-2 relative">
-              <Search className="pointer-events-none absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground" />
-              <Input
-                value={searchQuery}
-                onChange={(event) => setSearchQuery(event.target.value)}
-                placeholder="Search company directory by name or email..."
-                className="h-9 rounded-lg border-border bg-background pl-9 text-sm font-normal"
-                disabled={isSubmitting}
-              />
-            </div>
-
-            {normalizedQuery ? (
-              <div className="mt-3 rounded-lg border border-border bg-background p-3 shadow-sm">
-                {isSearching ? (
-                  <div className="py-3 text-center">
-                    <p className="text-sm font-normal text-muted-foreground">Searching users...</p>
-                  </div>
-                ) : searchError ? (
-                  <div className="py-3 text-center">
-                    <p className="text-sm font-medium text-red-600">{searchError}</p>
-                  </div>
-                ) : directoryResults.length > 0 ? (
-                  <div className="space-y-2">
-                    {directoryResults.map((directoryUser) => (
-                      <div key={directoryUser.id} className="flex items-center justify-between gap-3">
-                        <div className="flex min-w-0 items-center gap-3">
-                          <Avatar className="size-7 rounded-md">
-                            <AvatarFallback className="rounded-md bg-muted text-xs font-semibold text-foreground">
-                              {getInitials(directoryUser.name)}
-                            </AvatarFallback>
-                          </Avatar>
-
-                          <div className="min-w-0">
-                            <p className="truncate text-sm font-semibold text-foreground">
-                              {directoryUser.name}
-                            </p>
-                            <p className="truncate text-xs font-normal text-muted-foreground">{directoryUser.email}</p>
-                          </div>
-                        </div>
-
-                        <Button
-                          type="button"
-                          variant="ghost"
-                          size="icon-xs"
-                          className="text-foreground hover:bg-muted"
-                          onClick={() => addUserToSelection(directoryUser)}
-                          disabled={isSubmitting}
-                        >
-                          <CirclePlus className="h-4 w-4" />
-                        </Button>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="py-3 text-center">
-                    <p className="text-base font-semibold text-foreground">No user found</p>
-                    <p className="mt-1 text-xs font-normal text-muted-foreground">
-                      Your search &quot;{searchQuery.trim()}&quot; did not match any users.
-                    </p>
-                  </div>
-                )}
-              </div>
-            ) : null}
-
-            <div className="mt-3 rounded-lg border border-border bg-muted/80 p-3">
-              {mappedSelection.length > 0 ? (
-                <div className="max-h-25 space-y-2 overflow-y-auto pr-1 [scrollbar-width:thin] [scrollbar-color:#64748b_transparent] [&::-webkit-scrollbar]:w-1.5 [&::-webkit-scrollbar-thumb]:rounded-full [&::-webkit-scrollbar-thumb]:bg-muted0">
-                  {mappedSelection.map((selection) => (
-                    <div key={selection.id} className="flex items-center justify-between gap-3">
-                      <div className="flex min-w-0 items-center gap-3">
-                        <Avatar className="size-7 rounded-md">
-                          <AvatarFallback className="rounded-md bg-muted text-xs font-semibold text-foreground">
-                            {getInitials(selection.name)}
-                          </AvatarFallback>
-                        </Avatar>
-
-                        <div className="min-w-0">
-                          <p className="truncate text-sm font-semibold text-foreground">{selection.name}</p>
-                          <p className="truncate text-xs font-normal text-muted-foreground">{selection.email}</p>
-                        </div>
-                      </div>
-
-                      <Button
-                        type="button"
-                        variant="ghost"
-                        size="icon-xs"
-                        className="text-red-400 hover:bg-red-50 hover:text-red-500"
-                        onClick={() => removeUserFromSelection(selection.id)}
-                        disabled={isSubmitting}
-                      >
-                        <Trash2 className="h-3.5 w-3.5" />
-                      </Button>
-                    </div>
-                  ))}
-                </div>
-              ) : (
-                <div className="py-2 text-center text-xs font-normal text-muted-foreground">
-                  No users selected for this role.
-                </div>
-              )}
-            </div>
-
-            {error ? <p className="mt-2 text-sm font-medium text-red-600">{error}</p> : null}
-
-            <div className="mt-3 flex items-center justify-end gap-2">
-              <Button
-                type="button"
-                variant="outline"
-                className="text-sm font-medium"
-                onClick={() => onOpenChange(false)}
-                disabled={isSubmitting}
-              >
-                Cancel
-              </Button>
-              <Button
-                type="button"
-                onClick={handleSubmit}
-                disabled={isSubmitting || mappedSelection.length === 0}
-                className="bg-primary px-4 text-sm font-medium text-primary-foreground hover:bg-primary/90"
-              >
-                {isSubmitting ? "Confirming..." : "Confirm Mapping"}
-              </Button>
-            </div>
-          </div>
-        ) : user ? (
-          <div className="space-y-4">
-            <div className="flex items-center justify-between">
-              <h3 className="text-lg font-semibold text-foreground">Change User Role</h3>
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-xs"
-                className="text-muted-foreground hover:bg-muted"
-                onClick={() => onOpenChange(false)}
-              >
-                <X className="h-3.5 w-3.5" />
-              </Button>
-            </div>
-
-            <div className="space-y-1">
-              <label htmlFor="user-role" className="text-sm font-medium text-foreground">
-                Role
-              </label>
-              <select
-                id="user-role"
-                value={selectedRole}
-                onChange={(event) => setSelectedRole(event.target.value as UserRole)}
-                disabled={isSubmitting || activeUser?.id === currentUserId}
-                className="flex h-10 w-full rounded-md border border-border bg-background px-3 text-sm font-normal text-foreground outline-none ring-offset-background focus-visible:ring-2 focus-visible:ring-border"
-              >
-                {ROLE_OPTIONS.map((option) => (
-                  <option key={option.value} value={option.value}>
-                    {option.label}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            {error ? <p className="text-sm font-medium text-red-600">{error}</p> : null}
-
-            <div className="flex flex-col items-end gap-2 pt-1">
-              <div className="flex w-full justify-end gap-2">
-                <Button
-                  type="button"
-                  variant="outline"
-                  className="text-sm font-medium"
-                  onClick={() => onOpenChange(false)}
-                  disabled={isSubmitting}
-                >
-                  Cancel
-                </Button>
-                <Button
-                  type="button"
-                  className="text-sm font-medium"
-                  onClick={handleSubmit}
-                  disabled={isSubmitting || !activeUser || activeUser.id === currentUserId}
-                >
-                  {isSubmitting ? "Updating..." : "Update Role"}
-                </Button>
-              </div>
-              
-              {activeUser?.id === currentUserId && (
-                <p className="text-right text-xs font-normal text-muted-foreground">
-                  You cannot modify your own role.
-                </p>
-              )}
-            </div>
-          </div>
-        ) : null}
-      </DialogContent>
-    </Dialog>
-  )
-}
